refactor(observability): fold pii_redaction into the package (#1677 orphan 1)
src/services/pii_redaction.py (124L Singapore-pattern PII detector) moves to
src/observability/pii_redaction.py per #1677's ruling — its only consumers
are observability's own audit helpers (now an intra-package import) and
extraction's CSV path (now via the published root; extraction already
declares observability in depends_on). detect_pii joins the published
language (__all__ + contract interface).

App-boundary baseline shrinks 59 -> 57 (both out:: edges to
src.services.pii_redaction die). The contract/readme "unregistered import"
confessions are updated — src.config is now the package's only remaining
unregistered import. common/advisor/todo.md's stale "move pii_redaction into
advisor/_guardrails" line (pre-#1677) is marked superseded: advisor's
_guardrails is the separate chat-stream redaction concern.

Closes nothing on its own — #1677's remaining orphans (assets.py,
annualized_income.py, test_account_purge.py) and the 46 inbound edges stay
open; annualized_income -> advisor is blocked on #1671's physical package.
