+++ docs/ssot/app-boundary-baseline.json
@@ -43,7 +43,6 @@
   "in::apps/backend/src/routers/workflow.py::src.platform.extension.workflow_events::list_workflow_events_response",
   "in::apps/backend/src/routers/workflow.py::src.platform.extension.workflow_events::update_workflow_event_status",
   "out::apps/backend/src/extraction/extension/_base.py::src.services::ai_streaming",
-  "out::apps/backend/src/extraction/extension/_csv.py::src.services.pii_redaction::detect_pii",
   "out::apps/backend/src/extraction/extension/brokerage_positions.py::src.services.assets::AssetService",
   "out::apps/backend/src/extraction/extension/review_queue.py::src.services.fx::FxRateError",
   "out::apps/backend/src/extraction/extension/review_queue.py::src.services.fx::get_exchange_rate",
@@ -53,7 +52,6 @@
   "out::apps/backend/src/extraction/extension/transaction_classification.py::src.services.ai_streaming::stream_ai_json",
   "out::apps/backend/src/ledger/extension/fx_revaluation.py::src.services.fx::FxRateError",
   "out::apps/backend/src/ledger/extension/fx_revaluation.py::src.services.fx::get_exchange_rate",
-  "out::apps/backend/src/observability/audit.py::src.services.pii_redaction::detect_pii",
   "out::apps/backend/src/reconciliation/extension/scoring.py::src.prompts.reconciliation::build_reconciliation_prompt",
   "out::apps/backend/src/reconciliation/extension/scoring.py::src.services.ai_streaming::AIStreamError",
   "out::apps/backend/src/reconciliation/extension/scoring.py::src.services.ai_streaming::accumulate_stream",
